// src/types/realEstate.ts
export type Property = {
	id: string;
	creationTimestamp: number;
	name: string;
	address: string;
	floor: number;
	code: string;
	wifiSsid: string;
	wifiPass: string;
	rooms: number;
	lockIn: string;
	lockOut: string;
	safe: string;
	insuranceName: string;
	insuranceDate: string;
	insuranceNumber: string;
	notes: string;
};

export type Tenant = {
	id: string;
	creationTimestamp: number;
	name: string;
	senderName: string;
	idNumber: string;
	nationalInsuranceNumber: string;
	birthDate: string;
	email: string;
	phone: string;
	address: string;
	roomId: string;
	apartmentId: string;
	rent: number;
	rentFirstMonth: string;
	rentFirstRent: number;
	deposit: number;
	account: string;
	accountDeposit: string;
	iceName: string;
	iceLastname: string;
	iceIdNumber: string;
	iceNationalInsuranceNumber: string;
	iceEmail: string;
	icePhone: string;
	iceAddress: string;
	insuranceName: string;
	insuranceNumber: string;
	insuranceDate: string;
	contractDate: string;
	contractDateStart: string;
	contractDateEnd: string;
	contractDateHandoff: string;
	notes: string;
	isContract: boolean;
	isDeposit: boolean;
	is1stRent: boolean;
	isInsurance: boolean;
	isWarranty: boolean;
	isKey: boolean;
	isProtocol: boolean;
	status: number;
};